+++ src/eca/features/prompt.clj
@@ -79,6 +79,10 @@
                             :file (if partial
                                     (format "<file partial=true path=\"%s\">...\n%s\n...</file>\n" path content)
                                     (format "<file path=\"%s\">%s</file>\n" path content))
+                            :agents-file (multi-str
+                                          (format "<agents-file description=\"Instructions following AGENTS.md spec.\" path=\"%s\">" path)
+                                          content
+                                          "</agents-file>\n")
                             :repoMap (format "<repoMap description=\"Workspaces structure in a tree view, spaces represent file hierarchy\" >%s</repoMap>\n" @repo-map*)
                             :cursor (format "<cursor description=\"User editor cursor position (line:character)\" path=\"%s\" start=\"%s\" end=\"%s\"/>\n"
                                             path